# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CMS KLY API",
        "description": "Documentation CMS KLY API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/setting-fe-menu/": {
            "get": {
                "tags": [
                    "Front End Menu"
                ],
                "summary": "Get Front End Menu",
                "operationId": "ca367025a5ae2bd4e63bb903644d6bed",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "401": {
                        "description": "unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The security token is invalid"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authentication_Token": []
                    }
                ]
            }
        },
...
        "/api/menu/": {
            "get": {
                "tags": [
                    "Menu"
                ],
                "summary": "Get Menu",
                "operationId": "77534129ea5818987994885726c819f6",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "401": {
                        "description": "unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The security token is invalid"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authentication_Token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Authentication_Token": {
                "type": "apiKey",
                "description": "An authorization header. Example: Token",
                "name": "token",
                "in": "query"
            }
        }
    }
}